vision-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
numpy==1.26.4
Pillow==10.4.0
onnxruntime==1.18.0
insightface==0.7.3
ultralytics==8.2.50
opencv-python==4.10.0.84
